chore: 컨트롤러명 수정, 폴더링 방식 변경
