+++ src/controller/beer/beerAllGET.js
@@ -1,0 +1,40 @@
+var express = require('express');
+var dbConObj = require('../../../config/db_info');
+var dbconn = dbConObj.init();
+
+const {
+  Beer,
+  Style_Small,
+  Aroma,
+  Country,
+  Select
+} = require('../../../models');
+
+const util = require('../../../modules/util');
+const statusCode = require('../../../modules/statusCode');
+const responseMessage = require('../../../modules/responseMessage');
+
+//const beerService = require("../service/beerService");
+
+module.exports = {
+  /* 전체 beer 불러오기 */
+  getAllBeer: async (req, res) => {
+    const cursor = req.body;
+
+    try {
+      const beers = await Beer.findAll({
+        attributes: [
+          'id', 'k_name', 'e_name', 'star_avg', 'thumbnail_image'
+        ],
+      });
+
+      const result = {};
+      result.beers = beers;
+      result.cursor = cursor;
+      return res.status(statusCode.OK).send(util.success(responseMessage.BEER_OK, result));
+    } catch (error) {
+      console.error(error);
+      return res.status(statusCode.INTERNAL_SERVER_ERROR).send(util.fail(statusCode.INTERNAL_SERVER_ERROR, responseMessage.BEER_READ_ALL_FAIL));
+    }
+  },
+};
