+++ src/routes/select.js
@@ -1,11 +1,12 @@
 const express = require('express');
+const selectPOST = require('../controller/select/selectPOST');
+const selectUPDATE = require('../controller/select/selectUPDATE');
 const router = express.Router();
-const selectController = require('../controller/selectController');
 
 /* GET users listing. */
-router.get('/', selectController.getCheckSelect);
-router.post('/', selectController.postFirstSelect);
-router.put('/style', selectController.modifySelectStyle);
-router.put('/aroma', selectController.modifySelectAroma);
+router.get('/', selectPOST.getCheckSelect);
+router.post('/', selectPOST.postFirstSelect);
+router.put('/style', selectUPDATE.modifySelectStyle);
+router.put('/aroma', selectUPDATE.modifySelectAroma);
 
 module.exports = router;
